requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
streamlit==1.28.1
pandas==2.1.3
requests==2.31.0
groq==0.4.1
pytest==7.4.3
httpx==0.25.1
